refactor: FavoriteRepository の責務は、ArticleRepository だったので favorite メソッドを移動
