+++ src/main/kotlin/com/example/realworldkotlinspringbootjdbc/usecase/favorite/FavoriteUseCase.kt
@@ -6,8 +6,8 @@ import arrow.core.Either.Right
 import arrow.core.Invalid
 import arrow.core.Valid
 import arrow.core.left
+import com.example.realworldkotlinspringbootjdbc.domain.ArticleRepository
 import com.example.realworldkotlinspringbootjdbc.domain.CreatedArticle
-import com.example.realworldkotlinspringbootjdbc.domain.FavoriteRepository
 import com.example.realworldkotlinspringbootjdbc.domain.RegisteredUser
 import com.example.realworldkotlinspringbootjdbc.domain.article.Slug
 import com.example.realworldkotlinspringbootjdbc.util.MyError
@@ -27,7 +27,7 @@ interface FavoriteUseCase {
 
 @Service
 class FavoriteUseCaseImpl(
-    val favoriteRepository: FavoriteRepository
+    val articleRepository: ArticleRepository
 ) : FavoriteUseCase {
     override fun execute(slug: String?, currentUser: RegisteredUser): Either<FavoriteUseCase.Error, CreatedArticle> {
         return when (val it = Slug.new(slug)) {
@@ -38,22 +38,22 @@ class FavoriteUseCaseImpl(
             /**
              * Slug が適切
              */
-            is Valid -> when (val favoriteResult = favoriteRepository.favorite(it.value, currentUser.userId)) {
+            is Valid -> when (val favoriteResult = articleRepository.favorite(it.value, currentUser.userId)) {
                 /**
-                 * お気に入り追加 成功
+                 * お気に入り追加 失敗
                  */
                 is Left -> when (favoriteResult.value) {
                     /**
                      * 原因: 作成済記事が見つからなかった
                      */
-                    is FavoriteRepository.FavoriteError.ArticleNotFoundBySlug -> TODO()
+                    is ArticleRepository.FavoriteError.ArticleNotFoundBySlug -> TODO()
                     /**
                      * 原因: 不明
                      */
-                    is FavoriteRepository.FavoriteError.Unexpected -> TODO()
+                    is ArticleRepository.FavoriteError.Unexpected -> TODO()
                 }
                 /**
-                 * お気に入り追加 失敗
+                 * お気に入り追加 成功
                  */
                 is Right -> favoriteResult
             }
